refactored, on goal touched scene no longer switches, instead the player is teleported to new location out of camera view.
