+++ trapped/Assets/Images/face.png.meta
