+++ trapped/Assets/Prefabs/Goal.prefab.meta
