+++ trapped/Assets/Prefabs/Player.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &4775421970329211983
+--- !u!1 &7047054490171775604
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,39 +8,40 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 4775421970329211971}
-  - component: {fileID: 4775421970329211968}
-  - component: {fileID: 4775421970329211969}
-  - component: {fileID: 4775421970329211982}
-  - component: {fileID: 114491349535480194}
-  m_Layer: 8
-  m_Name: OtherPlayer
+  - component: {fileID: 7047054490171775614}
+  - component: {fileID: 7047054490171775609}
+  - component: {fileID: 7047054490171775608}
+  - component: {fileID: 7047054490171775615}
+  - component: {fileID: 7047054490171775610}
+  - component: {fileID: 7047054490171775605}
+  m_Layer: 1
+  m_Name: Player
   m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4775421970329211971
+--- !u!4 &7047054490171775614
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4775421970329211983}
+  m_GameObject: {fileID: 7047054490171775604}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 3.52, y: 0, z: 0}
-  m_LocalScale: {x: 0.350075, y: 0.350075, z: 0.350075}
+  m_LocalPosition: {x: -9.82, y: -0.49, z: 0}
+  m_LocalScale: {x: 0.35007498, y: 0.35007498, z: 0.35007498}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &4775421970329211968
+--- !u!212 &7047054490171775609
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4775421970329211983}
+  m_GameObject: {fileID: 7047054490171775604}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -72,7 +73,7 @@ SpriteRenderer:
   m_SortingLayer: 2
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300000, guid: 42921b99523d86646a0b2ff8b26c0bde, type: 3}
-  m_Color: {r: 0.1981132, g: 0.3689199, b: 1, a: 1}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
@@ -82,14 +83,14 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!50 &4775421970329211969
+--- !u!50 &7047054490171775608
 Rigidbody2D:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4775421970329211983}
+  m_GameObject: {fileID: 7047054490171775604}
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
@@ -103,13 +104,13 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
---- !u!60 &4775421970329211982
+--- !u!60 &7047054490171775615
 PolygonCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4775421970329211983}
+  m_GameObject: {fileID: 7047054490171775604}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
@@ -132,15 +133,31 @@ PolygonCollider2D:
       - {x: -0.93, y: -0.95}
       - {x: 0.93, y: -0.95}
       - {x: 0.93, y: 0.95}
---- !u!114 &114491349535480194
+--- !u!114 &7047054490171775610
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4775421970329211983}
+  m_GameObject: {fileID: 7047054490171775604}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 42e210958985dac43b71b06599469f65, type: 3}
+  m_Script: {fileID: 11500000, guid: c32c6bde3b61bcc4886eb7ead9d6e42a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  moveSpeed: 5
+  movement: {x: 0, y: 0}
+--- !u!114 &7047054490171775605
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7047054490171775604}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d7bf49740706e334aaf730b974abdb29, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  otherPlayerObject: {fileID: 4775421970329211983, guid: 416583d299cb10342b6d73819d2a839f,
+    type: 3}
