+++ trapped/Assets/Prefabs/Player.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: c4b838ea8ec6a4e4298df04b1230250f
-DefaultImporter:
+guid: 51b3d930401eb3f459937fdc166e51dc
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
