--- trapped/Assets/SceneSwitch.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4f196e2926d38246a33bd4819b14ebe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
