+++ trapped/Assets/Scripts/GlobalBehavior.cs.meta
