+++ trapped/Assets/Scripts/GoalBehavior.cs.meta
