+++ trapped/Assets/Scripts/OtherPlayerBehavior.cs
