+++ trapped/ProjectSettings/TagManager.asset
@@ -14,7 +14,7 @@ TagManager:
   - UI
   - 
   - 
-  - 
+  - Player
   - 
   - 
   -
